+++ resources/views/components/input/datepicker.blade.php
@@ -17,7 +17,7 @@
             {{-- control only can still include icons, (no errors) --}}
             <x-gotime::input.control-datepicker {{ $attributes }} />
         @else
-            <x-gotime::v2.input.layout-control-group>
+            <x-gt-control-group>
                 <x-gotime::input.control-datepicker {{ $attributes }} />
-            </x-gotime::v2.input.layout-control-group>
+            </x-gt-control-group>
         @endif