+++ resources/views/components/input/number.blade.php
@@ -3,7 +3,7 @@
 @if($controlOnly)
     <x-gotime::input.layout-control-only {{ $attributes->merge(['type' => 'number', 'step' => 'any']) }} />
 @else
-    <x-gotime::v2.input.layout-control-group>
+    <x-gt-control-group>
         <x-gotime::input.layout-control-only {{ $attributes->merge(['type' => 'number', 'step' => 'any']) }} />
-    </x-gotime::v2.input.layout-control-group>
+    </x-gt-control-group>
 @endif