+++ resources/views/components/v2/input/input.blade.php
@@ -3,7 +3,7 @@
 @if($controlOnly)
     <x-gotime::v2.input.control-input {{ $attributes }} />
 @else
-    <x-gotime::v2.input.layout-control-group>
+    <x-gt-control-group>
         <x-gotime::v2.input.control-input {{ $attributes }} />
-    </x-gotime::v2.input.layout-control-group>
+    </x-gt-control-group>
 @endif