+++ CHANGELOG.md
@@ -14,6 +14,7 @@
     - `emit()`, and `dispatchBrowserEvent()` to `dispatch()`
 - Updated: pagination layout
 - Updated: return value to float for `MoneyCast`
+- Added: Form controls and partials
 
 ## 0.15.0
 - Moved: layouts to components directory